% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertMe.R
\name{convertMe}
\alias{convertMe}
\title{A function to calculate the percent of methylation and convert methylkit
objects to GRanges objects. Note that it does not support methylDiff objects.}
\usage{
convertMe(WGBS_data)
}
\arguments{
\item{WGBS_data}{WGBS data in one of the following methylKit objects:
methylRawList, methylRaw, methylRawListDB,methylBaseDB, methylBase.}
}
\value{
a GRanges object containing the methylation percent of samples
}
\description{
A function to calculate the percent of methylation and convert methylkit
objects to GRanges objects. Note that it does not support methylDiff objects.
}
\examples{
library(methylKit)
data("methylKit") # Load the data
basegr <- convertMe(methylBase.obj)
rawlistgr <- convertMe(methylRawList.obj)
}
