% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findSignatures.R
\name{findSignatures}
\alias{findSignatures}
\title{A function to construct a signature matrix}
\usage{
findSignatures(samples, sampleMeta, atlas = NULL)
}
\arguments{
\item{samples}{dataframe, has first column "CpGs", rest of columns are samples (must have column name as sample accession ID which should be found in sampleMeta), rows are CpGs}

\item{sampleMeta}{dataframe, must have Experiment_accession for accession ID of each sample, and Biosample_term_name for cell type of sample, rows are samples}

\item{atlas}{dataframe, the reference atlas to which new signatures can be added, , if not present then a new reference atlas will be created using sample(s)}
}
\value{
A dataframe filteredExtendedMethAtlas which contains all cell types in atlas (if given), and those in samples added by cell type, has first column "CpGs", rest of columns are cell types, rows are CpGs, cells are methylation values
}
\description{
A function to construct a signature matrix
}
\examples{
 exampleAtlas_1 = readRDS(system.file("reference_atlas_nodup.RDS", package = "deconvR"))
 exampleMeta_1  = data.table("Experiment_accession" = as.character(1:25), "Biosample_term_name" = "example_cell_type")
 colnames(exampleAtlas_1)[2:26] =  as.character(1:25)
 findSignatures(samples = exampleAtlas_1, sampleMeta = exampleMeta_1)

 exampleAtlas_2 = readRDS(system.file("reference_atlas_nodup.RDS", package = "deconvR"))
 exampleMeta_2  = data.table("Experiment_accession" = as.character(1:25), "Biosample_term_name" = colnames(exampleAtlas_2)[-1])
 exampleSamples_2 = exampleAtlas_2[,c(1,26)]
 exampleAtlas_2 = exampleAtlas_2[,c(1:25)]
 exampleMeta_2    = exampleMeta_2[25,]
 colnames(exampleSamples_2)[2]   = "25"
 findSignatures(samples = exampleSamples_2, sampleMeta = exampleMeta_2, atlas = exampleAtlas_2)

}
\keyword{"CpGs",}
\keyword{CpGs}
\keyword{are}
\keyword{cell}
\keyword{column}
\keyword{columns}
\keyword{first}
\keyword{of}
\keyword{rest}
\keyword{rows}
\keyword{signaturehas}
\keyword{types,}
