% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IlluminaMethEpicB5ProbeIDs}
\alias{IlluminaMethEpicB5ProbeIDs}
\title{A dataset Illumina probe IDs of 400000 genomic loci (identified using the
“seqnames”, “ranges”, and “strand” values).}
\format{
GRanges object with 400000 ranges and 1 metadata column:
\describe{
\item{metadata}{contains seqnames ranges strand and ID}
...
}
}
\source{
\url{https://support.illumina.com/downloads/infinium-methylationepic-v1-0-product-files.html}
}
\usage{
data(IlluminaMethEpicB5ProbeIDs)
}
\description{
A dataset Illumina probe IDs of 400000 genomic loci (identified using the
“seqnames”, “ranges”, and “strand” values).
}
\keyword{datasets}
