% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeBigTable.R
\name{makeBigTable}
\alias{makeBigTable}
\title{A function to generate a dataframe of mixed cell-type origin simulated samples using given reference atlas.}
\usage{
makeBigTable(numberOfSamples, reference)
}
\arguments{
\item{numberOfSamples}{The number of simulated samples to be generated in the dataframe.}

\item{reference}{A reference dataframe with which to generate the simulations. Should be dataframe with first column "CpGs" for Illumina IDs, and rest of columns = cell types. Should not have duplicate Illumina IDs.}
}
\value{
A dataframe "bigExampleTable" which contains mixed cell-type origin simulated samples. First column "CpGs" for Illumina IDs, and rest of columns = cell types.
}
\description{
A function to generate a dataframe of mixed cell-type origin simulated samples using given reference atlas.
}
\keyword{simulation}
