% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reference_atlas}
\alias{reference_atlas}
\title{The comprehensive human methylome reference atlas
This object contains the methylation values of 11715 genomic loci
(identified using the “seqnames”, “ranges”, and “strand” values) for a
single example sample. It can be used directly as WGBS_data in BSmeth2Probe.}
\format{
data.frame object with 6000 CpG loci and 25 human cell types column:
\describe{
\item{cell type columns}{For each cell type and CpG locus, a methylation
valuebetween 0 and 1 is provided. This value represents the fraction of
methylated bases of the CpG locus.}
\item{IDs}{CpG loci IDs for each cell type.}
...
}
}
\source{
\url{https://doi.org/10.1038/s41467-018-07466-6}
}
\usage{
data("reference_atlas")
}
\description{
The comprehensive human methylome reference atlas
This object contains the methylation values of 11715 genomic loci
(identified using the “seqnames”, “ranges”, and “strand” values) for a
single example sample. It can be used directly as WGBS_data in BSmeth2Probe.
}
\references{
Moss, J. et al.  (2018). Comprehensive human cell-type
methylation atlas reveals origins of circulating cell-free DNA in health
and disease. Nature communications, 9(1), 1-12.
}
\keyword{datasets}
