% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateCellMix.R
\name{simulateCellMix}
\alias{simulateCellMix}
\title{A function to generate a dataframe of mixed cell-type origin simulated samples using given reference atlas.}
\usage{
simulateCellMix(
  numberOfSamples,
  mixingVector = NULL,
  reference = readRDS(system.file("reference_atlas_nodup.RDS", package = "deconvR"))
)
}
\arguments{
\item{numberOfSamples}{The number of simulated samples to be generated in the dataframe.}

\item{mixingVector}{Specify the cell origin proportions. If numberOfSamples = 1, this can be a vector of length = number of cell types in reference.
Otherwise, this is a dataframe with rows for cell types (must be equal to cell types in reference) and columns for samples.
Cells contain the proportion of the sample from the cell type. Use zeros for any unused cell type. If this object is not given,
will use random values for the simulation.}

\item{reference}{A dataframe containing CpG signatures of different cell types used to generate the simulation. The first column should contain a unique ID
(e.g. target ID) which can be used in deconvolution to match rows of the reference to rows of the bulk. All subsequent columns are cell types.
Rows are CpGs. Each cell contains the methylation value for the cell type at the CpG location. If not given, defaults to a reference atlas which is included in this
package (see deconvR/inst/reference_atlas_nodup.RDS). This atlas contains 25 cell types, 6105 CpGs (and their target IDs), and therefore 152,625 methylation values.}
}
\value{
A list containing two data frames.
First: A dataframe which contains mixed cell-type origin simulated samples. The first column contains a unique ID (used from reference) which can be used in
deconvolution to match rows of the reference to rows of the bulk. All subsequent columns are cell types. Rows are CpGs. Each cell contains the methylation
value for the cell type at the CpG location.
Second: A dataframe with the cell proportions of the generated samples. Each row is a sample. Columns are cell types.
}
\description{
A function to generate a dataframe of mixed cell-type origin simulated samples using given reference atlas.
}
\examples{
simulateCellMix(50)
simulateCellMix(numberOfSamples=100,
  reference=readRDS(system.file("reference_atlas_nodup.RDS", package = "deconvR")))
simulateCellMix(1,mixingVector = c(0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0))
simulateCellMix(1,data.frame( c(0,0,0,0,0,0,0,0.5,0,0,0,0,0,0,0,0,0,0,0,0.5,0,0,0,0,0)))
simulateCellMix(2,data.frame( c(0,0,0,0,0,0,0,0.5,0,0,0,0,0,0,0,0,0,0,0,0.5,0,0,0,0,0),
  c(0.1,0,0,0,0,0,0.5,0,0,0.2,0,0,0,0,0,0,0.2,0,0,0,0,0,0,0,0)))
}
\keyword{simulation}
